requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
printjson==1.0.1
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
sqlparse==0.5.3
urllib3==2.3.0
